Fix incorrect leech threshold setting saving

+++ ios/Settings.swift
@@ -206,7 +206,7 @@ protocol SettingProtocol {
   @Setting(true, #keyPath(showPreviousLevelGraph)) static var showPreviousLevelGraph: Bool
   @Setting(true, #keyPath(showKanaOnlyVocab)) static var showKanaOnlyVocab: Bool
 
-  @Setting(1.0, #keyPath(leechThreshold)) static var leechThreshold: Double
+  @Setting(1.0, #keyPath(leechThreshold)) static var leechThreshold: Float
 
   @Setting(false, #keyPath(seenFullAnswerPrompt)) static var seenFullAnswerPrompt: Bool
